chore: reorganize and rename LotCom.DataAccess namespace to LotCom.Database
