--- LotCom/DataAccess/Enums/AppEnvironment.cs
@@ -1,7 +1,0 @@
-namespace LotCom.DataAccess.Enums;
-
-public enum AppEnvironment
-{
-    DOTNET,
-    JavaScript
-}
