--- LotCom/DataAccess/Enums/AppName.cs
@@ -1,10 +1,0 @@
-namespace LotCom.DataAccess.Enums;
-
-public enum AppName
-{
-    Scanner,
-    Printer,
-    Watcher,
-    Client,
-    API
-}
