--- LotCom/DataAccess/Enums/AppPlatform.cs
@@ -1,7 +1,0 @@
-namespace LotCom.DataAccess.Enums;
-
-public enum AppPlatform
-{
-    Windows,
-    Cognex
-}
