--- LotCom/DataAccess/Extensions/AppEnvironmentExtensions.cs
@@ -1,18 +1,0 @@
-using LotCom.DataAccess.Enums;
-
-namespace LotCom.DataAccess.Extensions;
-
-public static class AppEnvironmentExtensions
-{
-    public static string ToString(AppEnvironment Value)
-    {
-        if (Value == AppEnvironment.JavaScript)
-        {
-            return "JavaScript";
-        }
-        else
-        {
-            return ".NET";
-        };
-    }
-}
