--- LotCom/DataAccess/Extensions/AppNameExtensions.cs
@@ -1,30 +1,0 @@
-using LotCom.DataAccess.Enums;
-
-namespace LotCom.DataAccess.Extensions;
-
-public static class AppNameExtensions
-{
-    public static string ToString(AppName Value)
-    {
-        if (Value == AppName.Scanner)
-        {
-            return "Scanner";
-        }
-        else if (Value == AppName.Printer)
-        {
-            return "Printer";
-        }
-        else if (Value == AppName.Watcher)
-        {
-            return "Watcher";
-        }
-        else if (Value == AppName.Client)
-        {
-            return "Client";
-        }
-        else
-        {
-            return "API";
-        }
-    }
-}
