+++ LotCom/Database/Enums/AppEnvironment.cs
@@ -1,0 +1,7 @@
+namespace LotCom.Database.Enums;
+
+public enum AppEnvironment
+{
+    DOTNET,
+    JavaScript
+}
