+++ LotCom/Database/Enums/AppName.cs
@@ -1,0 +1,10 @@
+namespace LotCom.Database.Enums;
+
+public enum AppName
+{
+    Scanner,
+    Printer,
+    Watcher,
+    Client,
+    API
+}
