+++ LotCom/Database/Enums/AppPlatform.cs
@@ -1,0 +1,7 @@
+namespace LotCom.Database.Enums;
+
+public enum AppPlatform
+{
+    Windows,
+    Cognex
+}
